fix(console): change to use working directory looking for `kics.config`
Closes #5242

This PR focus on changing setup for config file. KICS was looking for
config file on path it received on path flag (-p), the problem with this
approach is that was impossible to configure path flag on `kics.config`
file without using config flag (--config).

So, to solve this problem, the behavior is changed to look for config
file on the working directory, this should solve this issue and docker
issue too.

This has the same behavior as queries default path, as you can see on
`internal/console/helpers/helpers.go:133` and I think this is another
reason to change `kics.config` file looking to this behavior, since it
will standardize how KICS look for files and directories.

Docs was updated to fit with this new behavior.

Signed-off-by: Felipe Avelar <felipe.avelar@outlook.com>