feat: add Regal for linting Rego
Hello, friends at Kics! This PR introduces [Regal](https://github.com/styrainc/regal)
for linting the massive Rego library included in this project. Given the size of
this policy library, it should come as no surprise that a linter has much to say :)
I've tried to fix the reported issues that seem safe to address, while configuring
Regal to ignore the rest, leaving it to the maintainers of this project to decide
what they'll want to enforce going forward. I've annotated the provided config file
with some pointers and suggestions to try and make that chore a little easier.

Reported issues fixes as part of this:

Idiomatic:
* [custom-has-key-construct](https://docs.styra.com/regal/rules/idiomatic/custom-has-key-construct)
* [custom-in-construct](https://docs.styra.com/regal/rules/idiomatic/custom-in-construct)

Bugs:
* [rule-shadows-builtin](https://docs.styra.com/regal/rules/bugs/rule-shadows-builtin)

Style
* [default-over-else](https://docs.styra.com/regal/rules/style/default-over-else)
* [unconditional-assignment](https://docs.styra.com/regal/rules/style/unconditional-assignment)

Imports
* [redundant-alias](https://docs.styra.com/regal/rules/imports/redundant-alias)
* [import-after-rule](https://docs.styra.com/regal/rules/imports/import-after-rule)

I've also added linting of Rego as a new job for CI.

Let me know what you all think!

Signed-off-by: Anders Eknert <anders@styra.com>