fix #6997 - install jq on debian
fix #6997 - install jq on debian so that json results files can easily be queried inside the container. 

This should help alleviate the need for those users parsing the json results file to resort to grep/cut/awk and regex, where it is easy to make mistakes or commit obtuse code.